make ensureLocalAccountIndex run on every page, which makes it work with react-fast-refresh
This also requires a wrapper to router.push to be used
