+++ packages/webapp/src/pages/_app.tsx
@@ -8,15 +8,17 @@ import { ConnectKitProvider } from "connectkit"
 import { NextPage } from "next"
 import type { AppType } from "next/app"
 import Head from "next/head"
-import { WagmiConfig } from "wagmi"
+import { useAccount, WagmiConfig } from "wagmi"
 
-import { wagmiConfig } from "src/chain"
+import { ensureLocalAccountIndex, wagmiConfig } from "src/chain"
 import jotaiDebug from "src/components/lib/jotaiDebug"
 import { GlobalErrorModal } from "src/components/modals/globalErrorModal"
 import { useIsHydrated } from "src/hooks/useIsHydrated"
 import { useErrorConfig } from "src/store/hooks"
 
 import "src/styles/globals.css"
+import { useRouter } from "next/router"
+import { ComponentType, useEffect } from "react"
 
 // =================================================================================================
 
@@ -29,9 +31,6 @@ export type FablePage = NextPage<{ isHydrated: boolean }>
 // =================================================================================================
 
 const MyApp: AppType = ({ Component, pageProps }) => {
-  const errorConfig = useErrorConfig()
-  const isHydrated = useIsHydrated()
-
   return (
     <>
       <Head>
@@ -42,16 +41,56 @@ const MyApp: AppType = ({ Component, pageProps }) => {
       <WagmiConfig config={wagmiConfig}>
         <ConnectKitProvider>
           {jotaiDebug()}
-          <Component { ...pageProps } isHydrated={isHydrated} />
+          <ComponentWrapper Component={Component} pageProps={pageProps} />
         </ConnectKitProvider>
       </WagmiConfig>
-
-      {/* Global error modal for errors that don't have obvious in-flow resolutions. */}
-      {isHydrated && errorConfig && <GlobalErrorModal config={errorConfig} />}
     </>
   )
 }
 
 export default MyApp
+
+// =================================================================================================
+
+/**
+ * Wrapper for the main app component. This is necessary because we want to use the Wagmi account
+ * and the `useAccount` hook can only be used within a WagmiConfig.
+ */
+const ComponentWrapper = ({
+  Component,
+  pageProps
+}: {
+  Component: ComponentType
+  pageProps: any
+}) => {
+  const { address } = useAccount()
+  const isHydrated = useIsHydrated()
+  const errorConfig = useErrorConfig()
+
+  if (process.env.NODE_ENV === "development") { // constant
+    // eslint-disable-next-line react-hooks/rules-of-hooks
+    const router = useRouter()
+    const accountIndex = parseInt(router.query.index as string)
+    // eslint-disable-next-line react-hooks/rules-of-hooks
+    useEffect(() => {
+      if (accountIndex === undefined || isNaN(accountIndex)) return
+      if (accountIndex < 0 || 9 < accountIndex) return
+      void ensureLocalAccountIndex(accountIndex)
+    }, [accountIndex, address])
+
+    // It's necessary to update this on address, as Web3Modal (and possibly other wallet frameworks)
+    // will ignore our existence and try to override us with their own account (depending on how
+    // async code scheduling ends up working out).
+
+    // To carry the `index` query parameter to other parts of the app, be sure to use the `navigate`
+    // function from `utils/navigate.ts` instead of `router.push`.
+  }
+
+  return <>
+    <Component { ...pageProps } isHydrated={isHydrated} />
+    {/* Global error modal for errors that don't have obvious in-flow resolutions. */}
+    {isHydrated && errorConfig && <GlobalErrorModal config={errorConfig} />}
+  </>
+}
 
 // =================================================================================================
